Extract SqlKeywords utility class
